#768: automatically remove HTML `lang` attribute from `app.html` (#772)
