+++ .changeset/eight-oranges-float.md
@@ -1,0 +1,5 @@
+---
+"@inlang/sdk-js": minor
+---
+
+automatically remove HTML `lang` attribute from app template
