+++ .changeset/hip-jokes-run.md
@@ -1,0 +1,5 @@
+---
+"@inlang/sdk-js": minor
+---
+
+read folder structure of project from `svelte.config.js`
