+++ source-code/sdk-js/src/adapter-sveltekit/examples/test-app/languages/de.json
@@ -1,3 +1,3 @@
 {
-	"welcome": "Willkommen bei inlang"
+  "welcome": "Willkommen bei inlang"
 }
