+++ source-code/sdk-js/src/adapter-sveltekit/examples/test-app/languages/en.json
@@ -1,3 +1,3 @@
 {
-	"welcome": "Welcome to inlang"
+  "welcome": "Welcome to inlang"
 }
